Fixed some warnings
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6443 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/strlist.cpp
@@ -114,7 +114,7 @@ unsigned int ExeHeadStringList::find(const void *ptr, unsigned int cchF, WORD co
     if (processed)
     {
       char *pTmp = convert_processed_string_to_ansi(bufMB,find,codepage);
-      cbMB = pTmp ? pTmp - bufMB : 0;
+      cbMB = (int)(pTmp ? pTmp - bufMB : 0);
     }
     else
     {
